Add general left-right margin
Margin actually maked with padding as the
background color has to be applied. Call to action
section made with actual margin, as the background
color does not have to be applied in all the
available space, but the text inside has to have
actual padding to separate the content from the
borders.

Note: This could be done in a more clear way,
using the margin property to make the actual
margin and using more divs to make it structurally
possible.